+++ pygeos/test/test_creation.py
@@ -70,17 +70,21 @@ def test_linestrings_from_xyz():
     assert str(actual) == "LINESTRING Z (0 2 0, 1 3 0)"
 
 
+def test_linestrings_invalid_shape_scalar():
+    with pytest.raises(ValueError):
+        pygeos.linestrings((1, 1))
+
+
 @pytest.mark.parametrize(
     "shape",
     [
         (2, 1, 2),  # 2 linestrings of 1 2D point
         (1, 1, 2),  # 1 linestring of 1 2D point
         (1, 2),  # 1 linestring of 1 2D point (scalar)
-        (2,),  # 1 2D point (scalar)
     ],
 )
 def test_linestrings_invalid_shape(shape):
-    with pytest.raises(ValueError):
+    with pytest.raises(pygeos.GEOSException):
         pygeos.linestrings(np.ones(shape))
 
 
@@ -99,6 +103,11 @@ def test_linearrings_unclosed():
     assert str(actual) == "LINEARRING (1 0, 1 1, 0 1, 0 0, 1 0)"
 
 
+def test_linearrings_invalid_shape_scalar():
+    with pytest.raises(ValueError):
+        pygeos.linearrings((1, 1))
+
+
 @pytest.mark.parametrize(
     "shape",
     [
@@ -111,19 +120,24 @@ def test_linearrings_unclosed():
         (2, 3, 2),  # 2 linearrings of 3 2D points
         (1, 3, 2),  # 1 linearring of 3 2D points
         (3, 2),  # 1 linearring of 3 2D points (scalar)
-        (2,),  # 1 2D point (scalar)
     ],
 )
 def test_linearrings_invalid_shape(shape):
     coords = np.ones(shape)
-    with pytest.raises(ValueError):
+    with pytest.raises(pygeos.GEOSException):
         pygeos.linearrings(coords)
 
     # make sure the first coordinate != second coordinate
     coords[..., 1] += 1
-    with pytest.raises(ValueError):
+    with pytest.raises(pygeos.GEOSException):
         pygeos.linearrings(coords)
 
+
+def test_linearrings_all_nan():
+    coords = np.full((4, 2), np.nan)
+    with pytest.raises(pygeos.GEOSException):
+        pygeos.linearrings(coords)
+    
 
 def test_polygon_from_linearring():
     actual = pygeos.polygons(pygeos.linearrings(box_tpl(0, 0, 1, 1)))
@@ -175,6 +189,11 @@ def test_2_polygons_with_different_holes():
     assert pygeos.area(actual).tolist() == [96.0, 24.0]
 
 
+def test_polygons_not_enough_points_in_shell_scalar():
+    with pytest.raises(ValueError):
+        pygeos.polygons((1, 1))
+
+
 @pytest.mark.parametrize(
     "shape",
     [
@@ -187,18 +206,22 @@ def test_2_polygons_with_different_holes():
         (2, 3, 2),  # 2 linearrings of 3 2D points
         (1, 3, 2),  # 1 linearring of 3 2D points
         (3, 2),  # 1 linearring of 3 2D points (scalar)
-        (2,),  # 1 2D point (scalar)
     ],
 )
 def test_polygons_not_enough_points_in_shell(shape):
     coords = np.ones(shape)
-    with pytest.raises(ValueError):
+    with pytest.raises(pygeos.GEOSException):
         pygeos.polygons(coords)
 
     # make sure the first coordinate != second coordinate
     coords[..., 1] += 1
-    with pytest.raises(ValueError):
+    with pytest.raises(pygeos.GEOSException):
         pygeos.polygons(coords)
+
+
+def test_polygons_not_enough_points_in_holes_scalar():
+    with pytest.raises(ValueError):
+        pygeos.polygons(np.ones((1, 4, 2)), (1, 1))
 
 
 @pytest.mark.parametrize(
@@ -213,17 +236,16 @@ def test_polygons_not_enough_points_in_shell(shape):
         (2, 3, 2),  # 2 linearrings of 3 2D points
         (1, 3, 2),  # 1 linearring of 3 2D points
         (3, 2),  # 1 linearring of 3 2D points (scalar)
-        (2,),  # 1 2D point (scalar)
     ],
 )
 def test_polygons_not_enough_points_in_holes(shape):
     coords = np.ones(shape)
-    with pytest.raises(ValueError):
+    with pytest.raises(pygeos.GEOSException):
         pygeos.polygons(np.ones((1, 4, 2)), coords)
 
     # make sure the first coordinate != second coordinate
     coords[..., 1] += 1
-    with pytest.raises(ValueError):
+    with pytest.raises(pygeos.GEOSException):
         pygeos.polygons(np.ones((1, 4, 2)), coords)